+++ src/navigation/include/Pathfinding.hpp
@@ -31,21 +31,21 @@ private:
 	point_t						_buoyPos;
 	size_t						_buoyGraphIndex;
 
-	std::vector<rectangle_t>			parseObstacles();
-	rectangle_t							parseBoundingBox(std::string const & strBoundingBox);
-	void								generateObstaclesGraph();
-	void 								addObstacleAdjList(rectangle_t const & obstacle, size_t index);
-	void								generateAdjList(rectangle_t const & lhs, rectangle_t const & rhs);
-	void								generateNodeAdjList(point_t nodePos, size_t nodeIndex, Graph& graph);
-	std::vector<std::pair<int, double>>	djikstra(int start, int end, Graph const & graph);
-	std::vector<point_t>				convertNodeToPoint(std::list<int> nodePath);
-	static bool							isVisited(int index, std::vector<int> const & visited);
-	static std::pair<int, double>		getMinNode(std::vector<std::pair<int, double>> const & path, std::vector<int> const & visited);
-	static double						calculateDist(point_t a, point_t b);
-	static bool							isHitObstacle(point_t const start, point_t const end, rectangle_t const & obstacle);
-	static bool							isHitItself(point_t const start, point_t const end, rectangle_t const & obstacle, size_t index);
-	static bool							isIntersect(point_t p1, point_t q1, point_t p2, point_t q2);
-	static bool							areRectangleEqual(rectangle_t const & lhs, rectangle_t const & rhs);
+	std::vector<rectangle_t>				parseObstacles();
+	rectangle_t								parseBoundingBox(std::string const & strBoundingBox);
+	void									generateObstaclesGraph();
+	void 									addObstacleAdjList(rectangle_t const & obstacle, size_t index);
+	void									generateAdjList(rectangle_t const & lhs, rectangle_t const & rhs);
+	void									generateNodeAdjList(point_t nodePos, size_t nodeIndex, Graph& graph);
+	std::vector<std::pair<size_t, double>>	djikstra(size_t start, size_t end, Graph const & graph);
+	std::vector<point_t>					convertNodeToPoint(std::list<size_t> nodePath);
+	static bool								isVisited(size_t index, std::vector<size_t> const & visited);
+	static std::pair<size_t, double>		getMinNode(std::vector<std::pair<size_t, double>> const & path, std::vector<size_t> const & visited);
+	static double							calculateDist(point_t a, point_t b);
+	static bool								isHitObstacle(point_t const start, point_t const end, rectangle_t const & obstacle);
+	static bool								isHitItself(point_t const start, point_t const end, rectangle_t const & obstacle, size_t index);
+	static bool								isIntersect(point_t p1, point_t q1, point_t p2, point_t q2);
+	static bool								areRectangleEqual(rectangle_t const & lhs, rectangle_t const & rhs);
 
 public:
 	Pathfinding();
